salesAgent/requirements.txt
livekit-agents
livekit-agents[cartesia,silero,openai]
